passing double as long and lat to preserve precision

+++ app/src/main/java/com/brandonhogan/weathertracker/api/DarkSkyAPI.java
@@ -10,6 +10,6 @@ public interface DarkSkyAPI {
 
     @GET("{latitude},{longitude}?exclude=minutely&units=si")
     Observable<DarkSkyResponse> getForecast(
-                                            @Path(value = "latitude", encoded = true) long latitude,
-                                            @Path(value = "longitude", encoded = true) long longitude);
+                                            @Path(value = "latitude", encoded = true) double latitude,
+                                            @Path(value = "longitude", encoded = true) double longitude);
 }
